requirements.txt
Flask~=1.1.2
PyMySQL~=0.10.1
Werkzeug~=1.0.1
PyJWT~=1.7.1